move type SelectOption and update imports
